milestones/VDM/3.1/sth-cloud/sequences/audio-training-seq/requirements.txt
pyee
tensorflow==2.13.0 
tensorflow-io==0.33.0 
numpy
scipy
noisereduce
scramjet-framework-py